# Optimize logging collapse results