--- a/api/mod_ws_integration_test.go
+++ b/api/mod_ws_integration_test.go
@@ -41,6 +41,7 @@ func TestPollLog(t *testing.T) {
 	gateway := New()
 	gateway.setup()
 
+	log.ClearWriters()
 	log.SetWriter(log.LoggerWebsocket, gateway)
 
 	keys, err := skademlia.NewKeys(1, 1)