--- a/log/mod.go
+++ b/log/mod.go
@@ -27,7 +27,8 @@ import (
 
 var (
 	output = &multiWriter{
-		writers: make(map[string]io.Writer),
+		writers:        make(map[string]io.Writer),
+		writersModules: make(map[string]map[string]struct{}),
 	}
 	logger = zerolog.New(output).With().Timestamp().Logger()
 
@@ -95,7 +96,13 @@ func SetLevel(ls string) {
 }
 
 func SetWriter(key string, writer io.Writer) {
-	output.SetWriter(key, writer)
+	var modules []string
+	if cw, ok := writer.(ConsoleWriter); ok {
+		for k := range cw.FilteredModules {
+			modules = append(modules, k)
+		}
+	}
+	output.SetWriter(key, writer, modules...)
 }
 
 func Node() zerolog.Logger {
@@ -132,4 +139,10 @@ func Sync(event string) zerolog.Logger {
 
 func Metrics() zerolog.Logger {
 	return metrics
+}
+
+// Write to only the writers that have filter for the module.
+func Write(module string, msg []byte) error {
+	_, err := output.WriteFilter(msg, module)
+	return err
 }