--- a/log_test.go
+++ b/log_test.go
@@ -1,3 +1,5 @@
+// +build !integration,unit
+
 package wavelet
 
 import (