--- a/log_test.go
+++ b/log_test.go
@@ -17,6 +17,7 @@ import (
 
 func TestCollapseResultsLogger(t *testing.T) {
 	log.ClearWriters()
+	defer log.ClearWriters()
 
 	logger := NewCollapseResultsLogger()